+++ src/test/java/uk/gov/crowncommercial/dts/scale/cat/service/SupplierServiceTest.java
@@ -43,7 +43,7 @@ import uk.gov.crowncommercial.dts.scale.cat.repo.RetryableTendersDBDelegate;
 /**
  *
  */
-@SpringBootTest(classes = {SupplierService.class, RPAGenericService.class},
+@SpringBootTest(classes = {SupplierService.class},
     webEnvironment = WebEnvironment.NONE)
 @Slf4j
 @EnableConfigurationProperties({RPAAPIConfig.class})
